+++ src/form/form.tsx
@@ -1,20 +1,11 @@
-import React, { useRef, createContext } from 'react';
-// import Input from './InputField';
+import React, {
+    useRef,
+    useContext
+  } from "react";
 
-export const FormContext = createContext<ContextType>({ inputRefs: {}, errorRefs: {}, initialValues: {}});
+import { FormContext, FormContextType, GenericHTMLInput, ObjectType } from '../Forminho';
 
-export type initialValuesType = {
-    [key: string]: any
-}
 
-type FormHandlerHookType = {
-    onSubmitHandler: Function,
-    errors: { current: initialValuesType },
-    refs: { current: initialValuesType },
-    formRef: any,
-    contextValue: ContextType,
-    onLiveErrorFeedback?: Function
-}
 
 const getRefKeys = (refs: {[key:string]: HTMLInputElement}): Array<string> => Object.keys(refs).filter(key => key !== 'undefined');
 
@@ -27,18 +18,28 @@ const getRefKeys = (refs: {[key:string]: HTMLInputElement}): Array<string> => Ob
 //     }, {});
 //     return valuesFromRefs;
 // }
-export const handleFieldError = (contextValue: ContextType) => {
+
+
+export const handleFieldError = (contextValue: FormContextType) => {
     const setFieldError = (fieldName: string, errorMessage: string) => {
         // Check whether the field is on focus
         // If its not on focus then we dont set the errorText otherwise all fields could show errors before they were interacted with
-        const { current: currentElement } = contextValue.inputRefs[fieldName];
+        console.log('CONTEXT NO HANDLE FIELD', contextValue)
+        const { current: currentElement } = contextValue.fieldRefs[fieldName];
         const { current: currentErrorSpan } = contextValue.errorRefs[fieldName];
         console.log('CURRENT FIELD ', currentElement)
         // console.log('IS ON FOCUS? ', onfocus)
-        if (document.activeElement === currentElement) currentErrorSpan.innerText = errorMessage;
+        if (document.activeElement === currentElement) {
+            if (currentErrorSpan) {
+                currentErrorSpan.innerText = errorMessage;
+            }
+        }
     }
     const clearFieldError = (fieldName: string) => {
-        contextValue.errorRefs[fieldName].current.innerText = '';
+        if (contextValue.errorRefs[fieldName].current) {
+            const spanCurrent = contextValue.errorRefs[fieldName].current as HTMLSpanElement;
+            spanCurrent.innerText = '';
+        }
     }
 
     return {
@@ -47,111 +48,131 @@ export const handleFieldError = (contextValue: ContextType) => {
     }
 };
 
-
-const useForminhoHandler = ({onSubmitHandler, refs, formRef, contextValue, onLiveErrorFeedback}: FormHandlerHookType) => {
-    // const { setFieldError, clearFieldError } = handleFieldError(contextValue);
-
-    const onChangeHandler = (event: any) => {
-        console.log('CHANGE')
-        // console.log(event.target)
-        console.log(event.target.value)
-        console.log(event.target.name)
-
-        console.log('CONTEXT VALUES')
-        console.log(contextValue)
-
-        console.log('FORM VALUES')
-        const formRefValues = getValuesFromFormRef(formRef);
-        console.log(formRefValues)
-
-        if(onLiveErrorFeedback) onLiveErrorFeedback(formRefValues, contextValue);
-    };
-
-    const getValuesFromFormRef = (formRef: any) => {
-        const initialValuesKeys = getRefKeys(refs.current);
-        // console.log('KEYS ', initialValuesKeys)
-        const values = initialValuesKeys.reduce((valuesObj, key) => {
-            // console.log('KEY FORM REF', key, formRef.current[key]?.type, formRef.current[key])
-            const currentRef = formRef.current[key];
-            if (currentRef) {
-                // If its a radio or checkbox then it cant be accessed with .value just yet
-                if (NodeList.prototype.isPrototypeOf(currentRef)) {
-                    const elementsArray: Array<HTMLInputElement> = Array.from(currentRef);
-                    // console.log('elementsArray: ', elementsArray)
-                    const elementType = elementsArray[0].type;
-                    if (elementType === 'radio') {
-                        // console.log('radioValue: ', currentRef.value)
-                        valuesObj[key] = currentRef.value;
-                    } else if (elementType === 'checkbox') {
-                        const checked = elementsArray.filter((input: any) => input.checked);
-                        const checkedValues = checked.map((input: any) => input.value);
-                        // console.log('checkedValues: ', checkedValues)
-                        valuesObj[key] = checkedValues;
-                    }
-                } else if (currentRef.type === 'checkbox') {
-                    // If its not a node list of radios or checkboxes it can be an isolated checkbox
-                    // console.log('CAIU CHECKBOX ISOLADO ', currentRef.checked, currentRef.value)
-                    valuesObj[key] = currentRef.checked;
-                } else {
+const getValuesFromFormRef = (formRef: React.MutableRefObject<HTMLFormElement | null>, initialValues: ObjectType) => {
+    const initialValuesKeys = getRefKeys(initialValues);
+    // console.log('KEYS ', initialValuesKeys)
+    const values = initialValuesKeys.reduce((valuesObj, key) => {
+        // console.log('KEY FORM REF', key, formRef.current[key]?.type, formRef.current[key])
+        const { current } = formRef as React.MutableRefObject<HTMLFormElement>;
+        const currentRef = current[key];
+        if (currentRef) {
+            // If its a radio or checkbox then it cant be accessed with .value just yet
+            if (NodeList.prototype.isPrototypeOf(currentRef)) {
+                const elementsArray: Array<HTMLInputElement> = Array.from(currentRef);
+                // console.log('elementsArray: ', elementsArray)
+                const elementType = elementsArray[0].type;
+                if (elementType === 'radio') {
+                    // console.log('radioValue: ', currentRef.value)
                     valuesObj[key] = currentRef.value;
+                } else if (elementType === 'checkbox') {
+                    const checked = elementsArray.filter((input: any) => input.checked);
+                    const checkedValues = checked.map((input: any) => input.value);
+                    // console.log('checkedValues: ', checkedValues)
+                    valuesObj[key] = checkedValues;
                 }
+            } else if (currentRef.type === 'checkbox') {
+                // If its not a node list of radios or checkboxes it can be an isolated checkbox
+                // console.log('CAIU CHECKBOX ISOLADO ', currentRef.checked, currentRef.value)
+                valuesObj[key] = currentRef.checked;
+            } else {
+                valuesObj[key] = currentRef.value;
             }
-            return valuesObj;
-        }, {});
-        // console.log('VALUES FROM FORM REF ', values)
-        return values;
+        }
+        return valuesObj;
+    }, {});
+    // console.log('VALUES FROM FORM REF ', values)
+    return values;
+};
+
+
+
+const Form = (props:any) => {
+    const formRef = useRef(null);
+    const context = useContext(FormContext) as FormContextType;
+    // console.log("Context in the Form: ", context);
+    context.formRef = formRef;
+    context.initialValues = props.initialValues || {};
+    console.log("Context in the Form: ", context);
+  
+    const updateLiveValue = (fieldName: string) => {
+        console.log('Updating liveValues: ', context.liveValuesRefs[fieldName])
+        if (context.liveValuesRefs[fieldName]) {
+            console.log('TEM NO LIVE VALUE')
+            const liveValueSpan = context.liveValuesRefs[fieldName].current as HTMLSpanElement;
+            const fieldRefCurrent = context.fieldRefs[fieldName].current as GenericHTMLInput;
+            liveValueSpan.innerText = fieldRefCurrent.value;
+            console.log(liveValueSpan.innerText)
+        }
     };
+  
+    // const onChange = (event) => {
+    //   const name = event.target.name;
+    // //   console.log("FORM ONCHANGE ---> ", name);
+    // //   console.log("CONTEXT IN FORM: ", context, context.formRef.current);
+    //   updateLiveValue(name);
+    // //   console.log(
+    // //     "Updated context values: ",
+    // //     context.liveValues[name].current.innerText,
+    // //     context.fieldRefs[name].current.value
+    // //   );
+  
+    // //   console.log("CALLING PROPS ONCHANGE");
+    //   props.onChange(event);
+    // };
+  
+    // useEffect(() => {
+    //   console.log("Form Effect -> ", formRef);
+    // });
 
-    const submitHandler = (event: any) => {
-        event.preventDefault();
-        console.log('SUBMIT')
-        const formRefValues = getValuesFromFormRef(formRef);
-        onSubmitHandler(formRefValues);
-    }
-
-    return {
-        onChangeHandler,
-        submitHandler
-    };
-}
-
-interface FormProps extends FormHandlerHookType {
-    initialValues: initialValuesType,
-    children: Array<React.ReactElement>
-}
-
-const Form = ({initialValues, onSubmitHandler, children, onLiveErrorFeedback}: FormProps) => {
-    const refs = useRef(initialValues);
-    const formRef = useRef<HTMLFormElement>(null);
+    // const formRef = useRef(null);
     // console.log('initialValues ', initialValues)
     // console.log(Object.keys(initialValues))
-    const initialErrors = Object.keys(initialValues).reduce((errorsObj, key) => {
-        // console.log('KEY ', key, errorsObj)
-        errorsObj[key] = null;
-        return errorsObj;
-    }, {});
-    const errors = useRef(initialErrors);
-    // console.log(initialValues)
-    // console.log('INITIAL REFS', refs)
-    const contextValue = {
-        inputRefs: {},
-        errorRefs: {},
-        initialValues
+    // const initialErrors = Object.keys(initialValues).reduce((errorsObj, key) => {
+    //     // console.log('KEY ', key, errorsObj)
+    //     errorsObj[key] = null;
+    //     return errorsObj;
+    // }, {});
+    // const errors = useRef(initialErrors);
+
+    // const formHandler = useForminhoHandler({onSubmitHandler, errors, refs, formRef, contextValue, onLiveErrorFeedback, setCurrentValues});
+
+    const onChange = (event: any) => {
+        console.log('CHANGE')
+        console.log(event.target.value)
+        const { name } = event.target;
+        console.log(name)
+
+        updateLiveValue(name);
+
+        const formRefValues = getValuesFromFormRef(context.formRef, context.initialValues);
+        console.log(formRefValues)
+
+        console.log('Updating context.currentValues')
+        context.currentValues = formRefValues;
+        console.log(context)
+
+        if(props.onLiveErrorFeedback) props.onLiveErrorFeedback(formRefValues, context);
+        console.log('CALLING ONCHANGE FROM THE SIGNUP')
+        if(props.onChangeHandler) props.onChangeHandler(event);
     };
-    const formHandler = useForminhoHandler({onSubmitHandler, errors, refs, formRef, contextValue, onLiveErrorFeedback});
 
-    console.log('Rendering')
-    console.log(errors)
-    console.log(refs)
-    console.log(formRef)
 
+    const onSubmit = (event: any) => {
+        event.preventDefault();
+        console.log('SUBMIT')
+        const formRefValues = getValuesFromFormRef(formRef, context.initialValues);
+        props.onSubmitHandler(formRefValues);
+    }
+
+
+  
+    console.log("REF DO FORM: ", formRef);
     return (
-        <FormContext.Provider value={contextValue}>
-            <form ref={formRef} className="main-form" onSubmit={formHandler.submitHandler} onChange={formHandler.onChangeHandler}>
-            {children}
-        </form>
-        </FormContext.Provider>
-    )
-};
+      <form ref={formRef} onSubmit={onSubmit} onChange={onChange}>
+        {props.children}
+        {/* <button>Send</button> */}
+      </form>
+    );
+  };
 
 export default Form;
